Renamed two modules, fixed tests
